Create custom API rather than reusing CDATA
This commit introduces a custom extension to the document API rather
than trying to re-use the existing XML CDATA API.

After considering it further, using an existing API in a similar but
somewhat different way could be an uncanny valley that hurts more than
it hinders.

Instead, I have introduced the ability to create `RawHTMLSection` nodes,
which are similar to CDATA in that they contain arbitrary, unparsed
HTML data, but make the intent more clear.