Add Options for trimming rightspace and stripping comments to serializer
Why: Because minifying html after serialization is very expensive,
especially when not all types of minifications are desired.  Having the
ability to do simple whitespace and comment stripping during
serialization stage simplifies this, and default false options means the
change is backwards compatible

How: Add an optional `opts` arg to the Serializer constructor and check
it in methods that output text

Also: Cleaned up toLowerCase method.